Make promiscuos strategy open channels to any interesting node.
Before, the promiscuous channel would simply take a random channel, and
evaluate it for quality. If the threshold was above a level it would
auto open.

The problem is, if it hadn't seen the channel before, the quality is not
good enough. Therefore to open a channel, it must be randomly selected
twice. When there are a lot of channels open, this takes a while.

To avoid this, and as a naive solution, appropriate for our promiscuous
strategy, we now auto open to any interesting node at the beginning of
the tick, before also adding new random channels to evaluate.