Reduce hash calls in hashedSecret
Fixes #1202

Because hashedSecret wasn't actually an instance, and reloaded values
every time it was `check()`ed, `findPreImage` was called whenever we
initialized, which was incorrectly assumed to be a no-op.

In core we use initialize to try and initialize on each tick, so we
initialize as soon as possible, therefore findPreImage was being called
repeatedly.

Instead of simply hacking this init call, this PR refactors hashedSecret
to actually store information in it's instance.