Delay Docker image builds after version commits
Previously Docker images would be built on the commit which also was
used as the basis for a new version. Therefore, if changes included
type differences a race-condition could occur which would break the
Docker build.

This change makes sure the Docker images are only build for the commits
which introduce the new version. At that points all dependencies are
aligned, NPM packages published and the Docker images can be build.