Simplify Docker image build and use npm packages instead of source (#1854)
* Delay Docker image builds after version commits

Previously Docker images would be built on the commit which also was
used as the basis for a new version. Therefore, if changes included
type differences a race-condition could occur which would break the
Docker build.

This change makes sure the Docker images are only build for the commits
which introduce the new version. At that points all dependencies are
aligned, NPM packages published and the Docker images can be build.

* Change Docker image generation to use npm packages

* Make Docker image builds on GCP work with new image

* Fix typo

* Make process in Dockerfile work again

* Fix debug log setup