Merge kiautschou into master (#1945)
* Simplify Docker image build and use npm packages instead of source (#1854)

* Delay Docker image builds after version commits

Previously Docker images would be built on the commit which also was
used as the basis for a new version. Therefore, if changes included
type differences a race-condition could occur which would break the
Docker build.

This change makes sure the Docker images are only build for the commits
which introduce the new version. At that points all dependencies are
aligned, NPM packages published and the Docker images can be build.

* Change Docker image generation to use npm packages

* Make Docker image builds on GCP work with new image

* Fix typo

* Make process in Dockerfile work again

* Fix debug log setup

* Fix path to cloudbuild configuration

Improve Github deploy workflow so it can be tested locally as well.

* Make ignored path more specific

* chore(release): publish v1.72.13

* Avado CI: new release

* Fix entrypoint in Avado docker image

This fixes a crash where the entrypoint module could not be found.
Also changed the tag of the docker image so that Avado only uses images
from release branches and not pre-releases.

* Remove tagging latest on any change

* chore(release): publish v1.72.14

* Avado CI: new release

Co-authored-by: HOPR CI robot <noreply@hoprnet.org>