Add basic Swagger-based Rest API v2
As part of this change the Rest API v1 has been migrated from restana
to express, just like the healthcheck APIs from hoprd and
cover-traffic-daemon. This was done for easier integration of Swagger
tools in the future.

The feature set of Rest API v2 is minimal in this change and will be
extended later. Focus is on the tooling and making sure tests keep
working.