connect: Add support for handling multiple protocols
We previously called all libp2p connection handlers with a single
specified and supported protocol-prefix which was tied to the version
the node is running. This prevented nodes with different versions
communicating within the same environment even though they might be
supporting that as per the protocol config.

This change adds support for providing multiple supported protocols to
libp2p, which will handle the negotiation and pick the right protocol.

The supported protocols are determined based on the environments from the
protocol config.