Incremental improvement of the naive implementation
Intermediate implementation which significantly reduced polling and made
the Mixer object a more async implementation. Basic sync APIs of the
Mixer are retained and it is possible to perform operations on the
struct in a synchronous fashion. The used data structure in the form of
PriorityQueue turned out to be a defect, as it is impossible to store
the same packet multiple times.