Remove pending balance
This PR removes the pending balance from the database.
It therefore allows the ticket issuer to issue more tickets than the current channel balance allows.
It is up to the ticket receiver to validate that it does not try to aggregate more than the channel balance allows. The ticket aggregator also performs this check.

The ticket redemption check on balance is done only on-chain.

Closes #5689