Migrate the indexer
Migrate indexer implementation from TS to Rust.

Restructure and modify the code to be compliant with the custom written
RPC provider wrapper returning a stream of Blocks.

Streamify operations, including the handler based operations, so that
the indexer operation can be effectively performed as a separate process
working over a stream in a sequential order.