Improve ChannelStatus logic
This PR changes the `current_timestamp()` platform type to return `SystemTime` instance.
The `ChannelStatus` enum is changed, so that the `PendingToClose` variant must be accompanied by the `SystemTime` marking the channel closure timestamp.
Therefore the `closure_time` field is removed from the `ChannelEntry`.

The `ChannelClosureFinalizer` strategy is set to ignore `PendingToClose` channels that are overdue with closure over 1 day.