Fix duplicate strategy configuration from CLI
This PR fixes incorrectly doubled AutoRedeeming strategy configuration when the `auto_redeem_tickets` CLI argument was specified. This fixes the doubled on-chain transaction seen when the strategy was triggered.

Also when `default_strategy` is given on the CLI argument, it will be the only one used, clearing all the defaults.

Fixes #6015