Soft-restart indexer loop on error, expose `max_block_range` on the CLI
This PR exposes the `max_block_range` parameter on the CLI and ENV variables.

The default value of `max_block_range` is decreased from 2000 to 1000.

Also, as a follow-up to #6057, the logic in RPC log fetching loop (used by the Indexer) is improved, so it does not always panic on fatal errors.

This PR supersedes #6052
Refs #6039