Better handling of missing block errors
The log pagination algorithm in `ethers` might request blocks outside the fixed block range, which is handled differently by different RPC providers.

This PR handles the missing block error (-32001) so it is not considered fatal and further blocks are not requested.

Also the minimum block limit to use pagination has been increased to 100, and the maximum block range was as well increased to 2000.