Allow RPC provider to feed past blocks
In certain situations, the RPC provider can be feeding us blocks that were already processed. Previously we caused panic on this behavior. This PR changes this behavior, so that the node waits until RPC "upsyncs" with what we have in the DB currently. However, if the difference is more than 50 blocks, we still cause panic, as before.

This approach can be further improved, so that also the parent block hash is checked after "upsync", so that we know that this was not a chain re-org. In the case of re-org we should panic and ask user for a resync).