Path finding should only consider nodes that proven to be reliable
Previously, the assumption was that a new node (or a node that we don't have quality info about) is automatically considered reliable. This is not correct.

This PR considers all nodes (node, or missing) without a "quality" information to be automatically unreliable and thus not considered for automatic path finding.