Unify updating channels in the DB via an API to allow caching
This commit includes a refactoring on how channels are handled within the contract event handlers. It replaces the usage of the direct model manipulation with the newly introduced ChannelEditor API which streamlines modification and ensures validation rules are adhered to. Also, some unnecessary imports were removed and minor updates were made on the comments for better code readability.