Add tcpdump support to EchoServer for packet capture
This commit adds the ability to run tcpdump within the EchoServer class to capture packets during testing. The new feature is controlled by an optional parameter in the EchoServer constructor and is enabled in the test case by default. Logging and subprocess management have been implemented to handle tcpdump execution.