Fix TagBloomFilter behavior after updating the crate to 2.0
The `Bloom` filter no longer behaves as a ring buffer after the update to 2.0

This fact has been addressed in the `TagBloomFilter` wrapper, which makes sure that the count of the entries in the filter is correctly tracked.

Also, the false-positive rate decreased to 0.001%
and the capacity is kept at 10M replay tags.

Unit tests have been added to correctly test the new required behavior.

Benchmark was added for Bloom filter performance measurements