Early initialize lazy static metrics to prime them
Since all metrics are `lazy_static!`, it means thay are not primed (set to 0) immediately, but at a point they are first accessed.

This means, that they do not appear in the `node/metrics` endpoint until they were accessed for the first time.

This PR forcibly initializes the lazy static metrics everywhere, where they might not be used early.