Add loopback Session service for CT
The PR introduces a special target, which makes the Exit node to perform loopback of data received on a session back to it.

This is especially useful for the CT, so that no special echo service needs to be set up.

This loopback service has ID 0, thus for `target`, the user specifies `"target": {"Service": 0}` in the REST API endpoint (instead of usual ``"Plain": ...`)