Fix public key prefix bug
The protocol design was relying on the first byte of public key to never be equal to `0xff`. This was true for secp256k1 keys.

However, it is not the case for Ed25519 public keys as the top-most bit is set depending on the point sign.

This PR fixes this problem by re-organizing the header format, so that the path position is the first byte (possible values: `0x00`, `0x01`, `0x02`) and the `0xff` is set for the last hop.

Supersedes: #5665