Allow manual configuration of the ticket price
This change allows to configure manually the ticket price (using `protocol.outgoing_ticket_price` option) to override the value from the oracle with the higher possible price.

Prices lower than the network value are not possible and validated on the node start.

Also a bug has been fixed in the ticket validation, where the expected minimum ticket value should not be the value from the oracle, but it should be multiplied by the path position.