Cap on aggregation size, fix swarm notifiers
A cap of 10k tickets per aggregation batch was added. The maximum number of tickets in the aggregation request will not go over this limit.

This PR also fixes the notifier problem, that may occur when Arc<oneshot::Sender> was cloned by the cache. That resulted in the notification about ping/aggregation not being called and therefore, stalling tickets in the `BeingAggregated` state.