Make the allowlist of Session targets empty in default configs
Per default, nodes should not be acting as Exit nodes (receive Sessions), unless some list of allowed targets is specified.
This has been reflected in all default configs.
Note that this change does not affect Service-type Sessions, such as the ones used by the CT. These will always be allowed to be opened.

Also corrected other defaults in the config.