feat: invoke TBF saving via spawn_blocking
The Tag Bloom Filter (TBF) has a large size (~400MB) in memory, and therefore takes some time when it is being serialized to disk.

This PR fixes the problem where the save operation was invoked without `spawn_blocking`, possibly stalling the async-runtime.

In addition, `async_lock::RwLock` on the TBF is not needed, so it was replaced with a synchronous one (`std::sync::RwLock`), which is also more performant.